Init chat.js to create a socket on client side
Feat: Create socket connection on client side (front-end)

the socket on client side will connect to port 4000. All the input text
field values will be collected and when send button is clicked, the
chat event will be edited, gather and send the data including message
and name of the user to the server. The output section will display
sent message